batchrun: Convert logentry kind to integer
Change the type of the kind field of JobRunLogEntry model from string to
integer for improving query speed and storage requirements.

Update the log entry serializers so that the integer values (1 and 2)
will be serialized as "stdout" and "stderr" by setting "ints_as_names"
option to True.  This makes the API behave exactly in the same way as it
used to.

The included migration updates the field type from string to integer in
a few steps: It adds the integer field with name "kind2", copies data
from old kind to kind2, and finally replaces the old kind with kind2.
In addition it temporarily sets stdout as the default value for both
kind and kind2 in the beginning to allow much faster populating of the
kind2 field (or kind field in reverse migration), since most of the
entries (>99%) are stdout and the default allows them to be skipped on
the update.